fix: update liquidity swap improvements

+++ asp/src/api/handlers/burn.rs
@@ -150,7 +150,7 @@ pub async fn shielded_burn(
         let leaf_index = state.db.get_leaf_count()?;
         state
             .db
-            .insert_commitment(leaf_index as u32, &output0.commitment, Some(&tx_hash))?;
+            .insert_commitment(leaf_index, &output0.commitment, Some(&tx_hash))?;
         last_root = worker.insert_leaf(&output0.commitment).await?;
         tracing::debug!(leaf_index = leaf_index, "Inserted output_commitment_0");
     }
@@ -160,7 +160,7 @@ pub async fn shielded_burn(
         let leaf_index = state.db.get_leaf_count()?;
         state
             .db
-            .insert_commitment(leaf_index as u32, &output1.commitment, Some(&tx_hash))?;
+            .insert_commitment(leaf_index, &output1.commitment, Some(&tx_hash))?;
         last_root = worker.insert_leaf(&output1.commitment).await?;
         tracing::debug!(leaf_index = leaf_index, "Inserted output_commitment_1");
     }
@@ -170,7 +170,7 @@ pub async fn shielded_burn(
     // 10. Store the final root in DB (if we inserted anything)
     if !last_root.is_empty() {
         let new_count = state.db.get_leaf_count()?;
-        state.db.insert_root(&last_root, new_count as u32, Some(&tx_hash))?;
+        state.db.insert_root(&last_root, new_count, Some(&tx_hash))?;
 
         // 11. Submit the new Merkle root to Coordinator on-chain
         if let Some(ref relayer) = state.relayer {
@@ -192,7 +192,7 @@ pub async fn shielded_burn(
     let amount_0 = if amount_0_high == 0 {
         amount_0_low.to_string()
     } else {
-        format!("{}", (amount_0_high as u128).saturating_mul(u128::MAX).saturating_add(amount_0_low))
+        format!("{}", amount_0_high.saturating_mul(u128::MAX).saturating_add(amount_0_low))
     };
 
     let amount_1_low: u128 = req.output_note_1.amount_low.parse().unwrap_or(0);
@@ -200,7 +200,7 @@ pub async fn shielded_burn(
     let amount_1 = if amount_1_high == 0 {
         amount_1_low.to_string()
     } else {
-        format!("{}", (amount_1_high as u128).saturating_mul(u128::MAX).saturating_add(amount_1_low))
+        format!("{}", amount_1_high.saturating_mul(u128::MAX).saturating_add(amount_1_low))
     };
 
     Ok(Json(BurnResponse {
